Firmware/Liv_Flo_Sign/tools/EmojiDownloader/requirements.txt
pillow
beautifulsoup4
requests